src/autenticacion/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
python-dotenv==1.1.1
google-cloud-pubsub==2.23.1
debugpy==1.8.16
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
redis==5.0.1
pwdlib[argon2]==0.2.1
PyJWT==2.8.0